feat: 주문 수량에 따른 가격 표시 UI 반영
